Update Contributing to Obsidian.md
Missing a "c" on the word "can" under Translating the interface - added it in. Ironically, I came to fork and pull request a different typo - my recently downloaded Obsidian program shows "doesn't" instead of "don't" under Translating the docs, although it's corrected here. Is there a roving typo-maker conducting social engineering to engage nitpicky proofreaders? :D If so - it's working!