Clarified step-by-step Cloudflare instructions
Added clarifications on steps I got tripped up on a bit. Specifically, a) if setting up a subdomain, you still click through the root domain in Cloudflare, and b) not adding my personal Obsidian Publish url to the CNAME record. Also standardized on 'mysite.com' instead of a mix of 'my-site.com' and 'my-domain.com'.