Update Obsidian Sync and third-party services.md
I think the purpose of this article has shifted over time. The removed callout should be placed in a dedicated article on syncing Obsidian with third-party services, not an article about avoiding the use of Obsidian Sync with those services. (I just dealt with a case where the customer thought this meant it was okay, as long as it's configured appropriately).