Incorrect linking to advanced guide & core plugins
Noticed that Pane Layout guide isn't linking to any pages, seems to link to using tabs in obsidian online, so updated it to that. The embedded link to core plugins was incorrectly(?) trying to link to official plugins, which doesn't exist. Not sure if core plugins is wrong and should be updated in that page, or if the link has the wrong name.