Update Importer des données.md
- Translation and typo errors corrected ;
- Adding an external link to the Wikipedia page about personal knowledge management ;
- Adding new French translations ;
- Deleted non-pertinent information which are not in the VO doc ;
- Corrected bugs with external links.

I've added in this note the content of the one about importation from Roam Research. I've planned to add a translation for the documentation about each other software mentioned in the VO doc, but it'll take a lot of time, so for now I publish it like that and I'll add more later. I've made the choice to put all of this in one unique note with a little summary linking to the title for each software, but it can be discussed of course.